Moved common int code to utils class
