+++ lib/wab/impl.rb
@@ -13,6 +13,7 @@ require 'wab/impl/expr'
 require 'wab/impl/path_expr'
 require 'wab/impl/bool_expr'
 require 'wab/impl/shell'
+require 'wab/impl/utils'
 
 # Require the concrete Expr subclasses so a mapping table can be created for
 # the parser.
