+++ lib/wab/impl/configuration.rb
@@ -157,16 +157,16 @@ Bundler or directly, and try loading again.
           if node.is_a?(Hash)
             key = key.to_sym
             unless node.has_key?(key)
-              ai = key_to_int_ok(path[i + 1])
+              ai = Utils.key_to_int_ok(path[i + 1])
               node[key] = ai.nil? ? {} : []
             end
             node = node[key]
           elsif node.is_a?(Array)
-            key = key_to_int(key)
+            key = Utils.key_to_int(key)
             if key < node.length && -node.length < key
               node = node[key]
             else
-              ai = key_to_int_ok(path[i + 1])
+              ai = Utils.key_to_int_ok(path[i + 1])
               nn = ai.nil? ? {} : []
               if key < -node.length
                 node.unshift(nn)
@@ -184,7 +184,7 @@ Bundler or directly, and try loading again.
           key = key.to_sym
           node[key] = value
         elsif node.is_a?(Array)
-          key = key_to_int(key)
+          key = Utils.key_to_int(key)
           if key < -node.length
             node.unshift(value)
           else
@@ -226,32 +226,6 @@ Bundler or directly, and try loading again.
         node
       end
       alias [] get
-
-      def key_to_int(key)
-        return key if key.is_a?(Integer)
-
-        key = key.to_s if key.is_a?(Symbol)
-        if key.is_a?(String)
-          i = key.to_i
-          return i if i.to_s == key
-        end
-        return key if WAB::Utils.pre_24_fixnum?(key)
-
-        raise WAB::Error, 'path key must be an integer for an Array.'
-      end
-
-      # returns either an int or nil.
-      def key_to_int_ok(key)
-        return key if key.is_a?(Integer)
-
-        key = key.to_s if key.is_a?(Symbol)
-        if key.is_a?(String)
-          i = key.to_i
-          return i if i.to_s == key
-        end
-        return key if WAB::Utils.pre_24_fixnum?(key)
-        nil
-      end
 
     end # Configuration
   end # Impl
